NF: split Executor from MigrateUserData
