+++ AnkiDroid/src/test/java/com/ichi2/anki/servicelayer/scopedstorage/ExecutorTest.kt
@@ -16,7 +16,7 @@
 
 package com.ichi2.anki.servicelayer.scopedstorage
 
-import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.MigrateUserData.Executor
+import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.Executor
 import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.MigrationContext
 import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.Operation
 import org.hamcrest.MatcherAssert.assertThat
